feat(core-lib): move download function from console to core-lib & refactor export api

fix: fix minor thing
