+++ packages/core-lib/src/file-download/index.ts
@@ -1,0 +1,9 @@
+export const downloadByFileUrl = async (downloadUrl:string): Promise<void> => {
+    if (document) {
+        const link = document.createElement('a');
+        link.href = downloadUrl;
+        link.target = '_self';
+        link.click();
+        link.remove();
+    }
+};
